fix misplaced entry when opening day at end of month
Issue was caused when time was advanced by one day to search for day entry. The date was not reformatted after advancing the date by one so if the day was the end of the month then the day would be set to an invalid value. Fixed by calling mktime to reformat time to logical values